refactored releaseType, reset sort on create release,  and added loading to code generator
