+++ components/ReleaseSort/ReleaseSort.jsx
@@ -58,10 +58,13 @@ export default function ReleaseSort({ releases, onChange }) {
 
     onChange(sortedItems)
   }
+  useEffect(() => {
+    setSortBy("newest")
+  }, [releases])
 
   useEffect(() => {
     handleSort()
-  }, [sortBy, releases])
+  }, [sortBy])
 
   return (
     <div>
