+++ components/Releases/CreateRelease.jsx
@@ -8,20 +8,11 @@ import {
   DialogContent,
   DialogClose,
 } from "@/components/Dialog/Dialog"
-import IconMusicNotesPlus from "@/icons/music-notes-plus.svg"
 import Avatar from "../Avatar/Avatar"
 import PopoverTip from "../PopoverTip/PopoverTip"
 import { prependProtocol } from "@/utils/utils"
 import InputPasswordProtect from "../InputPasswordProtect/InputPasswordProtect"
-
-const releaseTypes = [
-  { id: 0, text: "Choose release type", isDisabled: true, isSelected: true },
-  { id: 1, text: "LP" },
-  { id: 2, text: "EP" },
-  { id: 3, text: "Single" },
-  { id: 4, text: "Compilation" },
-  { id: 5, text: "Soundtrack" },
-]
+import InputReleaseType from "../InputReleaseType/InputReleaseType"
 
 export default function CreateRelease({
   trigger,
@@ -264,27 +255,8 @@ export default function CreateRelease({
             onChange={(e) => setArtworkUrl(e.target.value)}
           />
           <p>Upload an image or paste an external link</p>*/}
-          <label className="label" htmlFor="type">
-            Type
-          </label>
-          <select
-            className="input select"
-            onChange={(e) => setType(e.target.value)}
-            id="type"
-            value={type}
-            required
-          >
-            {releaseTypes.map((releaseType) => (
-              <option
-                key={releaseType.id}
-                value={releaseType.value}
-                disabled={releaseType.isDisabled}
-                selected={releaseType.isSelected}
-              >
-                {releaseType.text}
-              </option>
-            ))}
-          </select>
+
+          <InputReleaseType type={type} setType={setType} />
           <p>You must include https:// in your links</p>
           <label className="label" htmlFor="yumUrl">
             Redemption (yum) Link
@@ -394,7 +366,7 @@ export default function CreateRelease({
             className="button"
             data-variant="primary"
             onClick={() => createNewRelease()}
-            disabled={!title || type == releaseTypes[5].text || noGO}
+            disabled={!title || type == "Choose release type" || noGO}
           >
             Create
           </button>
